Fix dtype issue in more robust way

+++ tests/test_env_checker.py
@@ -8,17 +8,17 @@ from stable_baselines3.common.env_checker import check_env
 
 class ActionDictTestEnv(gym.Env):
     action_space = Dict({"position": Discrete(1), "velocity": Discrete(1)})
-    observation_space = Box(low=-1.0, high=2.0, shape=(3,), dtype=np.float64)
+    observation_space = Box(low=-1.0, high=2.0, shape=(3,), dtype=np.float32)
 
     def step(self, action):
-        observation = np.array([1.0, 1.5, 0.5])
+        observation = np.array([1.0, 1.5, 0.5], dtype=self.observation_space.dtype)
         reward = 1
         done = True
         info = {}
         return observation, reward, done, info
 
     def reset(self):
-        return np.array([1.0, 1.5, 0.5])
+        return np.array([1.0, 1.5, 0.5], dtype=self.observation_space.dtype)
 
     def render(self, mode="human"):
         pass
